requirements.txt
blitz-bayesian-pytorch==0.2.3
gym==0.12.1
matplotlib==3.1.0
shapely==1.7.0
streamlit==0.60.0
torch==1.4.0
torchvision==0.5.0
wget==3.2
